+++ App.js
@@ -1,9 +1,7 @@
 import { useState } from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Navigator from './routes/Homestack'
-
-import Button from './components/home/Button';
+import Navigator from './routes/NavStack'
 
 const Stack = createNativeStackNavigator();
