+++ screens/initWallet/InitWalletScreen.tsx
